fix: 에러처리, 리턴타입 수정/추가 및 관리자 계정 조회 기능 수정

+++ src/admin/admin.service.ts
@@ -7,16 +7,16 @@ import moment from 'moment';
 import randomToken from 'rand-token';
 import _ from 'lodash';
 import { Admin } from 'src/admin/entities/admin.entity';
-import { SigninAdminDto } from 'src/admin/dto/signin.admin.dto';
-import { SignupAdminResDto } from 'src/admin/dto/signup.admin.res.dto';
-import { SignupAdminReqDto } from 'src/admin/dto/signup.admin.req.dto';
 import { User } from 'src/auth/entities/user.entity';
 import { Collection } from 'src/collections/entities/collection.entity';
 import { Photospot } from 'src/photospot/entities/photospot.entity';
 import { Meetup } from 'src/meetups/entities/meetup.entity';
 import { Faq } from 'src/admin/entities/faq.entity';
-import { CreateAdminFaqDto } from './dto/create.admin.faq.dto';
-import { UpdateAdminFaqDto } from './dto/update.admin.faq.dto';
+import { SignupAdminResDto } from 'src/admin/dto/signup.admin.res.dto';
+import { SigninAdminDto } from 'src/admin/dto/signin.admin.dto';
+import { SignupAdminReqDto } from 'src/admin/dto/signup.admin.req.dto';
+import { CreateAdminFaqDto } from 'src/admin/dto/create.admin.faq.dto';
+import { UpdateAdminFaqDto } from 'src/admin/dto/update.admin.faq.dto';
 
 @Injectable()
 export class AdminService {
@@ -31,8 +31,25 @@ export class AdminService {
   ) {}
 
   // 관리자 관리
-  async getAdminsList(alias: string) {
-    return this.adminRepository.createQueryBuilder(alias);
+  async getAdminsList(keyword: string, p: number = 1): Promise<any> {
+    const adminList = this.adminRepository.createQueryBuilder('admin');
+    if (keyword) {
+      adminList.where('admin.account LIKE :keyword OR admin.responsibility LIKE :keyword', {
+        keyword: `%${keyword}%`,
+      });
+    }
+
+    const take = 6;
+    const page: number = (p as any) > 0 ? parseInt(p as any) : 1;
+    const total = await adminList.getCount();
+    adminList.skip((page - 1) * take).take(take);
+
+    return {
+      data: await adminList.getMany(),
+      total,
+      page,
+      lastPage: Math.ceil(total / take),
+    };
   }
 
   private async hashPassword(password: string): Promise<string> {
@@ -121,7 +138,7 @@ export class AdminService {
 
   async deleteAdmin(id: number) {
     await this.isMasterAdmin(id);
-    this.adminRepository.delete(id);
+    return this.adminRepository.delete(id);
   }
 
   private async isMasterAdmin(id: number) {
@@ -185,19 +202,18 @@ export class AdminService {
   }
 
   async deleteAdminCollection(id: number) {
-    try {
-      await this.adminCollectionsRepository.findOne({ where: { id } });
-      return this.adminCollectionsRepository.softDelete(id);
-    } catch (error) {
-      throw new BadRequestException();
+    await this.adminCollectionsRepository.findOne({ where: { id } });
+    if (_.isNil(id)) {
+      throw new NotFoundException('해당 콜렉션 게시물을 찾을 수 없습니다.');
     }
+    return this.adminCollectionsRepository.softDelete(id);
   }
 
   // 포토스팟 관리
-  async getAdminAllPhotospot(id: number): Promise<Photospot[]> {
+  async getAdminPhotospotList(id: number): Promise<Photospot[]> {
     const photospots = await this.adminPhotospotsRepository.find({ where: { id } });
     if (!photospots.length) {
-      throw new NotFoundException('해당 콜렉션을 찾을 수 없습니다.');
+      throw new NotFoundException('해당 포토스팟 목록을 찾을 수 없습니다.');
     }
     return photospots;
   }
@@ -205,13 +221,16 @@ export class AdminService {
   async getAdminPhotospot(photospotId: number): Promise<Photospot> {
     const photospot = await this.adminPhotospotsRepository.findOne({ where: { id: photospotId } });
     if (_.isNil(photospot)) {
-      throw new NotFoundException('해당 포토스팟을 찾을 수 없습니다.');
+      throw new NotFoundException('해당 포토스팟 게시물을 찾을 수 없습니다.');
     }
     return photospot;
   }
 
   async deleteAdminPhotospot(photospotId: number) {
     await this.getAdminPhotospot(photospotId);
+    if (_.isNil(photospotId)) {
+      throw new NotFoundException('해당 포토스팟 게시물을 찾을 수 없습니다.');
+    }
     this.adminPhotospotsRepository.softDelete(photospotId);
   }
 
@@ -238,12 +257,11 @@ export class AdminService {
   }
 
   async deleteAdminMeetup(id: number) {
-    try {
-      await this.adminMeetupsRepository.findOne({ where: { id } });
-      return this.adminMeetupsRepository.delete(id);
-    } catch (error) {
-      throw new BadRequestException();
+    await this.adminMeetupsRepository.findOne({ where: { id } });
+    if (_.isNil(id)) {
+      throw new NotFoundException('해당 모임 게시물을 찾을 수 없습니다.');
     }
+    return this.adminMeetupsRepository.delete(id);
   }
 
   // 자주찾는질문 관리
@@ -271,7 +289,7 @@ export class AdminService {
   async getAdminFaq(id: number): Promise<Faq> {
     const faq = await this.adminFaqRepository.findOne({ where: { id } });
     if (_.isNil(faq)) {
-      throw new NotFoundException('해당 게시물을 찾을 수 없습니다.');
+      throw new NotFoundException('해당 자주찾는질문 게시물을 찾을 수 없습니다.');
     }
     return faq;
   }
@@ -283,7 +301,7 @@ export class AdminService {
   async updateAdminFaq(updateAdminFaqtDto: UpdateAdminFaqDto, id: number) {
     await this.getAdminFaq(id);
     if (_.isNil(id)) {
-      throw new NotFoundException('해당 게시물을 찾을 수 없습니다.');
+      throw new NotFoundException('해당 자주찾는질문 게시물을 찾을 수 없습니다.');
     }
     return this.adminFaqRepository.update({ id }, updateAdminFaqtDto);
   }
@@ -291,7 +309,7 @@ export class AdminService {
   async deleteAdminFaq(id: number) {
     await this.getAdminFaq(id);
     if (_.isNil(id)) {
-      throw new NotFoundException('해당 게시물을 찾을 수 없습니다.');
+      throw new NotFoundException('해당 자주찾는질문 게시물을 찾을 수 없습니다.');
     }
     return this.adminFaqRepository.softDelete(id);
   }
